requirements.txt
hidapi==0.7.99.post21
Pillow==6.2.1
pyusb==1.0.2
